Add parser for uplink radio messages.
This parser can handle two kind of messages: generic command messages
and configuration ones. 
For commands, the format is:
<cmd> <param>
Where <cmd> is the command and <param> is a simple integer parameter.

In case of configuration messages, the format is:
cfg <module> <param1>=<val1> <param2>=<val2> ...

'cfg' is a header which tells the parser that this is configuration message, 
then follows the <module> which needs to be configured and then a list of
<param>=<val> pairs for the parameters to be set.
At the end of the list of parameters, if at least one parameter has been
found, the configuration of <module> is reloaded.